# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Appoint API",
    "description": "Appointment Management System API",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.appoint.local/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tokens generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/owner/business": {
      "post": {
        "tags": ["Business"],
        "summary": "Create business",
        "operationId": "createBusiness",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slug"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "description": { "type": "string" },
                  "timezone": { "type": "string", "default": "UTC" },
                  "isSingleStaff": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business created"
          }
        }
      },
      "get": {
        "tags": ["Business"],
        "summary": "Get owner businesses",
        "operationId": "getBusinesses",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of businesses"
          }
        }
      }
    },
    "/owner/business/{businessId}": {
      "put": {
        "tags": ["Business"],
        "summary": "Update business",
        "operationId": "updateBusiness",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "timezone": { "type": "string" },
                  "isSingleStaff": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Business updated"
          }
        }
      }
    },
    "/owner/business/public/{slug}": {
      "get": {
        "tags": ["Business"],
        "summary": "Get business by slug (public)",
        "operationId": "getBusinessBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Business details"
          }
        }
      }
    },
...
    "/owner/availability/{staffId}": {
      "put": {
        "tags": ["Availability"],
        "summary": "Set staff availability",
        "operationId": "setAvailability",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "staffId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rules"],
                "properties": {
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "dayOfWeek": { "type": "number", "minimum": 0, "maximum": 6 },
                        "startTime": { "type": "string" },
                        "endTime": { "type": "string" },
                        "isWorkingDay": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated"
          }
        }
      },
      "get": {
        "tags": ["Availability"],
        "summary": "Get staff availability",
        "operationId": "getAvailability",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "staffId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability rules"
          }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Create booking (public)",
        "operationId": "createBooking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["businessSlug", "serviceId", "startTimeUtc", "clientName", "clientEmail"],
                "properties": {
                  "businessSlug": { "type": "string" },
                  "serviceId": { "type": "string" },
                  "staffId": { "type": "string" },
                  "startTimeUtc": { "type": "string", "format": "date-time" },
                  "clientName": { "type": "string" },
                  "clientEmail": { "type": "string", "format": "email" },
                  "clientPhone": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          },
          "409": {
            "description": "Time slot not available"
          }
        }
      },
      "get": {
        "tags": ["Bookings"],
        "summary": "Get bookings",
        "operationId": "getBookings",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED"] }
          },
          {
            "name": "staffId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          }
        }
      }
    },
    "/bookings/{bookingId}": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Update booking",
        "operationId": "updateBooking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "businessId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["CONFIRMED", "CANCELLED", "COMPLETED"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking updated"
          }
        }
      }
    }
  }
}